% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.r
\name{item}
\alias{item}
\alias{qiita_delete_item}
\alias{qiita_delete_stock}
\alias{qiita_get_item}
\alias{qiita_get_stock}
\alias{qiita_patch_item}
\alias{qiita_post_item}
\alias{qiita_put_stock}
\title{Qiita Item API}
\usage{
qiita_get_item(item_id = NULL, tag_id = NULL, user_id = NULL,
  query = NULL, per_page = 100L, page_offset = 0L, page_limit = 1L)

qiita_post_item(title, body, tags = list(qiita_tag("R")), coediting = FALSE,
  private = FALSE, gist = FALSE, tweet = FALSE)

qiita_delete_item(item_id)

qiita_patch_item(item_id, title, body, tags = list(qiita_tag("R")),
  private = FALSE)

qiita_delete_stock(item_id)

qiita_get_stock(item_id = NULL, user_id = NULL, per_page = 100L,
  page_offset = 0L, page_limit = 1L)

qiita_put_stock(item_id)
}
\arguments{
\item{item_id}{item(article) ID}

\item{tag_id}{tag ID}

\item{user_id}{user ID}

\item{query}{query string}

\item{per_page}{number of items per one page}

\item{page_offset}{page offset}

\item{page_limit}{max number of pages to aquire.}

\item{title}{title of the item}

\item{body}{body of the item}

\item{tags}{tags of the item}

\item{coediting}{If \code{TRUE}, the post will be editable by team members.}

\item{private}{private or not}

\item{gist}{push to gist or not}

\item{tweet}{tweet or not}
}
\description{
Qiita Item API
}

