% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qiita.r
\docType{class}
\name{Qiita}
\alias{Qiita}
\title{An R6 Class for Qiita API}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Qiita
}
\description{
An R6 Class for Qiita API
}
\keyword{datasets}

