% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comment.r
\name{qiita_get_comment}
\alias{qiita_get_comment}
\title{Qiita Comment API}
\usage{
qiita_get_comment(url, token, comment_id = NULL, item_id = NULL,
  per_page = 100L, page_offset = 0L, page_limit = 1L)
}
\arguments{
\item{url}{URL of Qiita (If you're not using Qiita:Team, this should be "https://qiita.com")}

\item{token}{Qiita API token}

\item{comment_id}{comment ID}

\item{item_id}{article ID}
}
\description{
Qiita Comment API
}

