% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item.r
\name{qiita_get_item}
\alias{qiita_get_item}
\title{Qiita Item API}
\usage{
qiita_get_item(url, token, item_id = NULL, tag_id = NULL, user_id = NULL,
  query = NULL)
}
\description{
Qiita Item API
}

