% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/user.r
\name{qiita_get_stocker}
\alias{qiita_get_stocker}
\title{Qiita User API}
\usage{
qiita_get_stocker(url, token, item_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)
}
\arguments{
\item{url}{URL of Qiita (If you're not using Qiita:Team, this should be "https://qiita.com")}

\item{token}{Qiita API token}

\item{item_id}{item id}

\item{user_id}{user id}
}
\description{
Qiita User API
}

