% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tag.r
\name{qiita_get_tag}
\alias{qiita_get_tag}
\title{Qiita Tagging API}
\usage{
qiita_get_tag(url, token, tag_id = NULL, user_id = NULL, per_page = 100L,
  page_offset = 0L, page_limit = 1L)
}
\arguments{
\item{url}{URL of Qiita (If you're not using Qiita:Team, this should be "https://qiita.com")}

\item{token}{Qiita API token}

\item{tag_id}{tag id}

\item{user_id}{user id}
}
\description{
Qiita Tagging API
}

