% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.r
\name{qiita_tag}
\alias{qiita_tag}
\title{Utililies}
\usage{
qiita_tag(name, version = NULL)
}
\arguments{
\item{name}{tag name}

\item{version}{version}
}
\description{
Utililies
}
\examples{
tag(name = "R", version = ">3.1")
}

