% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiitr-package.R
\docType{package}
\name{qiitr-package}
\alias{qiitr}
\alias{qiitr-package}
\title{qiitr: R Interface to Qiita API}
\description{
Qiita is a technical knowledge sharing and collaboration platform for programmers.
  See <https://qiita.com/api/v2/docs> for more information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yutannihilation/qiitr}
  \item Report bugs at \url{https://github.com/yutannihilation/qiitr/issues}
}

}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})

}
\keyword{internal}
