% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.r
\name{tag}
\alias{qiita_delete_tagfollow}
\alias{qiita_get_following_tag}
\alias{qiita_get_tag}
\alias{qiita_get_tagfollow}
\alias{qiita_put_tagfollow}
\alias{tag}
\title{Qiita Tagging API}
\usage{
qiita_get_tag(url, token, tag_id = NULL, user_id = NULL, per_page = 100L,
  page_offset = 0L, page_limit = 1L)

qiita_get_following_tag(url, token, user_id, per_page = 100L,
  page_offset = 0L, page_limit = 1L)

qiita_delete_tagfollow(url, token, tag_id)

qiita_get_tagfollow(url, token, tag_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_put_tagfollow(url, token, tag_id)
}
\arguments{
\item{url}{URL of Qiita (If you're not using Qiita:Team, this should be "https://qiita.com")}

\item{token}{Qiita API token}

\item{tag_id}{tag ID}

\item{user_id}{user ID}

\item{per_page}{number of items per one page}

\item{page_offset}{page offset}

\item{page_limit}{max number of pages to aquire.}
}
\description{
Qiita Tagging API
}

