% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.r
\name{user}
\alias{qiita_delete_users_following}
\alias{qiita_get_authenticated_user}
\alias{qiita_get_followees}
\alias{qiita_get_followers}
\alias{qiita_get_stockers}
\alias{qiita_get_users}
\alias{qiita_get_users_following}
\alias{qiita_put_users_following}
\alias{user}
\title{Qiita User API}
\usage{
qiita_get_stockers(item_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_get_users(user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_get_followees(user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_get_followers(user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_delete_users_following(user_id)

qiita_get_users_following(user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_put_users_following(user_id)

qiita_get_authenticated_user()
}
\arguments{
\item{item_id}{item(article) ID}

\item{per_page}{number of items per one page}

\item{page_offset}{page offset}

\item{page_limit}{max number of pages to aquire.}

\item{user_id}{user ID}
}
\description{
Qiita User API
}

