% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.r
\name{comment}
\alias{comment}
\alias{qiita_delete_comment}
\alias{qiita_get_comment}
\alias{qiita_patch_comment}
\alias{qiita_post_comment}
\title{Qiita Comment API}
\usage{
qiita_get_comment(comment_id = NULL, item_id = NULL, per_page = 100L,
  page_offset = 0L, page_limit = 1L)

qiita_delete_comment(comment_id)

qiita_patch_comment(comment_id, body)

qiita_post_comment(item_id, body)
}
\arguments{
\item{comment_id}{comment ID}

\item{item_id}{item(article) ID}

\item{per_page}{number of items per one page}

\item{page_offset}{page offset}

\item{page_limit}{max number of pages to aquire.}

\item{body}{body of the item}
}
\description{
Qiita Comment API
}

