% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.r
\name{item}
\alias{item}
\alias{qiita_delete_items}
\alias{qiita_delete_stocks}
\alias{qiita_get_items}
\alias{qiita_get_stocks}
\alias{qiita_patch_items}
\alias{qiita_post_items}
\alias{qiita_put_stocks}
\title{Qiita Item (Article) API}
\usage{
qiita_get_items(item_id = NULL, tag_id = NULL, user_id = NULL,
  query = NULL, per_page = 100L, page_offset = 0L, page_limit = 1L)

qiita_post_items(title, body, tags = qiita_tag("R"), coediting = FALSE,
  private = FALSE, gist = FALSE, tweet = FALSE)

qiita_delete_items(item_id)

qiita_patch_items(item_id, title, body, tags = list(qiita_tag("R")),
  private = FALSE)

qiita_delete_stocks(item_id)

qiita_get_stocks(item_id = NULL, user_id = NULL, per_page = 100L,
  page_offset = 0L, page_limit = 1L)

qiita_put_stocks(item_id)
}
\arguments{
\item{item_id}{item(article) ID}

\item{tag_id}{tag ID}

\item{user_id}{user ID}

\item{query}{query string}

\item{per_page}{number of items per one page}

\item{page_offset}{page offset}

\item{page_limit}{max number of pages to aquire.}

\item{title}{title of the item}

\item{body}{body of the item}

\item{tags}{tags of the item}

\item{coediting}{If \code{TRUE}, the post will be editable by team members.}

\item{private}{private or not}

\item{gist}{push to gist or not}

\item{tweet}{tweet or not}
}
\description{
Qiita Item (Article) API
}

