% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Qiita.r
\docType{class}
\name{Qiita}
\alias{Qiita}
\title{An R6 Class for Qiita API}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x00000000167d2d88> 
 - attr(*, "name")= chr "Qiita_generator"
}}
\usage{
Qiita
}
\description{
An R6 Class for Qiita API
}
\keyword{datasets}

