% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{qiita_api}
\alias{qiita_api}
\title{Send A Request To Qiita API}
\usage{
qiita_api(verb, url, path, token, payload = NULL, query = NULL,
  per_page = 100L, page_offset = 0L, page_limit = 1L)
}
\arguments{
\item{verb}{method type (GET, POST, DELETE, etc..)}

\item{url}{URL of Qiita (If you're not using Qiita:Team, this should be "https://qiita.com")}

\item{path}{path of API}

\item{token}{Qiita API token}

\item{payload}{JSON payload}

\item{query}{URI quer}

\item{per_page}{number of items/tags/users/etc. in one page}

\item{page_offset}{page number to start}

\item{page_limit}{the number of pages where no more requests will be sent}
}
\description{
An \link[httr]{httr} Wrapper for Qiita API.
}

