% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.r
\name{qiita_api}
\alias{qiita_api}
\title{A httr Wrapper for Qiita API}
\usage{
qiita_api(verb, url, path, token, payload = NULL, query = NULL)
}
\arguments{
\item{verb}{method type (GET, POST, DELETE, etc..)}

\item{url}{URL of Qiita (If you're not using Qiita:Team, this should be "https://qiita.com")}

\item{path}{path of API}

\item{token}{Qiita API token}

\item{payload}{JSON payload}

\item{query}{URI quer}
}
\description{
A httr Wrapper for Qiita API
}

