% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{qiita_api}
\alias{qiita_api}
\title{Send A Request To Qiita API}
\usage{
qiita_api(verb, path, payload = NULL, query = NULL, per_page = 100L,
  page_offset = 0L, page_limit = 1L)
}
\arguments{
\item{verb}{Method type (e.g. GET, POST, DELETE).}

\item{path}{Path of API.}

\item{payload}{JSON payload.}

\item{query}{Query strings.}

\item{per_page}{Number of entries (e.g. items, tags, users) in one page.}

\item{page_offset}{Number of offset pages.}

\item{page_limit}{Max number of pages to get.}
}
\description{
An \link[httr]{httr} Wrapper for Qiita API.
}

