% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.r
\name{tag}
\alias{qiita_delete_tags_following}
\alias{qiita_get_tags}
\alias{qiita_get_tags_followed_by_user}
\alias{qiita_get_tags_following}
\alias{qiita_put_tags_following}
\alias{tag}
\title{Qiita Tag API}
\usage{
qiita_get_tags(url, token, tag_id = NULL, user_id = NULL, per_page = 100L,
  page_offset = 0L, page_limit = 1L)

qiita_get_tags_followed_by_user(url, token, user_id, per_page = 100L,
  page_offset = 0L, page_limit = 1L)

qiita_delete_tags_following(url, token, tag_id)

qiita_get_tags_following(url, token, tag_id, per_page = 100L,
  page_offset = 0L, page_limit = 1L)

qiita_put_tags_following(url, token, tag_id)
}
\arguments{
\item{url}{URL of Qiita (If you're not using Qiita:Team, this should be "https://qiita.com").}

\item{token}{Qiita API token.}

\item{tag_id}{Tag ID.}

\item{user_id}{User ID.}

\item{per_page}{Number of items per one page.}

\item{page_offset}{Page offset.}

\item{page_limit}{Max number of pages to aquire.}
}
\description{
Qiita Tag API
}

