% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.r
\name{user}
\alias{qiita_delete_userfollow}
\alias{qiita_get_authenticated_user}
\alias{qiita_get_followee}
\alias{qiita_get_follower}
\alias{qiita_get_stocker}
\alias{qiita_get_user}
\alias{qiita_get_userfollow}
\alias{qiita_put_userfollow}
\alias{user}
\title{Qiita User API}
\usage{
qiita_get_stocker(item_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_get_user(user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_get_followee(user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_get_follower(user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_delete_userfollow(user_id)

qiita_get_userfollow(user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_put_userfollow(user_id)

qiita_get_authenticated_user()
}
\arguments{
\item{item_id}{item(article) ID}

\item{per_page}{number of items per one page}

\item{page_offset}{page offset}

\item{page_limit}{max number of pages to aquire.}

\item{user_id}{user ID}
}
\description{
Qiita User API
}

