% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.r
\name{user}
\alias{qiita_delete_userfollow}
\alias{qiita_get_authenticated_user}
\alias{qiita_get_followee}
\alias{qiita_get_follower}
\alias{qiita_get_stocker}
\alias{qiita_get_user}
\alias{qiita_get_userfollow}
\alias{qiita_put_userfollow}
\alias{user}
\title{Qiita User API}
\usage{
qiita_get_stocker(url, token, item_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_get_user(url, token, user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_get_followee(url, token, user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_get_follower(url, token, user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_delete_userfollow(url, token, user_id)

qiita_get_userfollow(url, token, user_id, per_page = 100L, page_offset = 0L,
  page_limit = 1L)

qiita_put_userfollow(url, token, user_id)

qiita_get_authenticated_user(url, token)
}
\arguments{
\item{url}{URL of Qiita (If you're not using Qiita:Team, this should be "https://qiita.com")}

\item{token}{Qiita API token}

\item{item_id}{item(article) ID}

\item{per_page}{number of items per one page}

\item{page_offset}{page offset}

\item{page_limit}{max number of pages to aquire.}

\item{user_id}{user ID}
}
\description{
Qiita User API
}

