% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.r
\name{util}
\alias{qiita_payload}
\alias{qiita_tag}
\alias{util}
\title{Utililies}
\usage{
qiita_tag(name, version = NULL)

qiita_payload(body = NULL, title = NULL, tags = NULL, gist = NULL,
  private = NULL, tweet = NULL)
}
\arguments{
\item{name}{tag name}

\item{version}{version}

\item{body}{body}

\item{title}{title}

\item{tags}{tags}

\item{gist}{gist}

\item{private}{private}

\item{tweet}{tweet}
}
\description{
Utililies
}
\examples{
qiita_tag(name = "R", version = ">3.1")
}

